% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hent_data.R
\name{check_connect}
\alias{check_connect}
\title{Check connection
Function to check that a connection to data.ssb.no is able to be established}
\usage{
check_connect(url)
}
\arguments{
\item{url}{String url address for connection to check}
}
\value{
Nothing is returned but a error or warning message is return if no connection is available
}
\description{
Check connection
Function to check that a connection to data.ssb.no is able to be established
}
\keyword{internal}
