% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hent_data.R
\name{CheckDate}
\alias{CheckDate}
\title{Internal function to check date}
\usage{
CheckDate(date)
}
\arguments{
\item{date}{Date}
}
\description{
Internal function to check date
}
\keyword{internal}
