% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert_table.R
\name{ConvertTable}
\alias{ConvertTable}
\title{Code level convert table (internal function)}
\usage{
ConvertTable(
  klass_data,
  code = "code",
  parentcode = "parentCode",
  level = "level"
)
}
\arguments{
\item{klass_data}{Klass data frame to convert}

\item{code}{Name of variable containing code (Default = "code")}

\item{parentcode}{Name of variable of parent code (Default = "parentCode")}

\item{level}{Name of variable containing level (Default = "level")}
}
\value{
A dataframe
}
\description{
Code level convert table (internal function)
}
\keyword{internal}
