% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klass_list.R
\name{correspond_list}
\alias{correspond_list}
\alias{CorrespondList}
\title{Correspondence list
Print a list of correspondence tables for a given classification with source and target IDs}
\usage{
correspond_list(classification, date = NULL)

CorrespondList(klass, date = NULL)
}
\arguments{
\item{classification}{Classification number}

\item{date}{Date for classification (format = "YYYY-mm-dd"). Default is current date}

\item{klass}{Deprecated; use `classification` instead.}
}
\value{
Data frame with list of corrsepondence tables, source ID and target ID.
}
\description{
Correspondence list
Print a list of correspondence tables for a given classification with source and target IDs
}
\examples{
\donttest{
correspond_list("7")
}
}
