% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateKlass-graph-navigate.R
\name{count_neighbors}
\alias{count_neighbors}
\title{Count the neighbors of a node.}
\usage{
count_neighbors(graph, node, mode)
}
\arguments{
\item{graph}{A graph generated by \code{\link{klass_graph}}.}

\item{node}{A node as returned by \code{\link{klass_node}} or
\code{\link[igraph]{V}}.}

\item{mode}{Whether to query outgoing (\sQuote{out}), incoming
(\sQuote{in}) edges, or both types (\sQuote{all}). This is
ignored for undirected graphs.}
}
\value{
A numeric vector of length one giving the number of neighbors.
}
\description{
Count the neighbors of a node.
}
\keyword{internal}
