% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateKlass-graph-build.R
\name{find_dates}
\alias{find_dates}
\title{For a given Klass code, produce a table of dates describing the valid-from
and valid-to dates of all versions of the code}
\usage{
find_dates(code, api_alle, api_endringer)
}
\arguments{
\item{code}{A Klass code}

\item{api_alle}{A table of all codes in the classification. See example.}

\item{api_endringer}{A table of all changes in the classification See
example.}
}
\value{
A \code{data.frame} with number of rows equal to the number of
  variants of the combination of code and name, determined by the changes the
  code has been involved in. The \code{data.frame} has two columns:

  \itemize{
   \item{"validFrom"}
   \item{"validTo"}
  }
}
\description{
For a given Klass code, produce a table of dates describing the valid-from
and valid-to dates of all versions of the code
}
\keyword{internal}
