% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateKlass-graph-build.R
\name{find_name}
\alias{find_name}
\title{Find the name of a code valid at a specific date.}
\usage{
find_name(code, validFrom, validTo, api_alle)
}
\arguments{
\item{code}{A Klass code}

\item{validFrom}{The date the code is valid from, in \code{YYYY-MM-DD} format.}

\item{validTo}{The date the code is valid to, in \code{YYYY-MM-DD} format.}

\item{api_alle}{A table of all codes in the classification. See example.}
}
\value{
The name of the code.
}
\description{
Find the name of a code valid at a specific date.
}
\keyword{internal}
