% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateKlass-graph-build.R
\name{find_variant_to}
\alias{find_variant_to}
\title{Find the variant of a code corresponding to a change *to* a specific code.}
\usage{
find_variant_to(x, changeOccurred, variants)
}
\arguments{
\item{x}{The code that is being changed}

\item{changeOccurred}{The date the change occurred}

\item{variants}{The variants lookup-table.}
}
\value{
The variant corresponding to the code \code{x} at date
\code{changeOccurred}.
}
\description{
Find the variant of a code corresponding to a change *to* a specific code.
}
\seealso{
[find_variant_from()]
}
\keyword{internal}
