% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{formattering_kommune}
\alias{formattering_kommune}
\title{Format vector for kommune codes}
\usage{
formattering_kommune(x)
}
\arguments{
\item{x}{- vector of character type for kommune codes}
}
\value{
vector of character
}
\description{
Format vector for kommune codes
}
\keyword{internal}
