% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{formattering_nace}
\alias{formattering_nace}
\title{Format vector for industry codes}
\usage{
formattering_nace(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Formatted charcter vector
}
\description{
Format vector for industry codes
}
\keyword{internal}
