% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klass_list.R
\name{get_family}
\alias{get_family}
\alias{GetFamily}
\title{Identify corresponding family from a classification number}
\usage{
get_family(classification)

GetFamily(klass)
}
\arguments{
\item{classification}{Classification number}

\item{klass}{Deprecated; use `classification` instead.}
}
\value{
Family number
}
\description{
Identify corresponding family from a classification number
}
\examples{
get_family(classification = 7)
}
