% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klass_list.R
\name{get_name}
\alias{get_name}
\alias{GetName}
\title{Get the name of a classification version}
\usage{
get_name(version)

GetName(version)
}
\arguments{
\item{version}{Version number}
}
\value{
string or vector of strings with name of version
}
\description{
Get the name of a classification version
}
\examples{
get_name("33")
}
