% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hent_data.R
\name{get_variant_name}
\alias{get_variant_name}
\title{Get variant name
Internal function for fetching the variant name based on the number}
\usage{
get_variant_name(variant)
}
\arguments{
\item{variant}{The variant number}
}
\description{
Get variant name
Internal function for fetching the variant name based on the number
}
\keyword{internal}
