% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klass_list.R
\name{get_version}
\alias{get_version}
\alias{GetVersion}
\title{Get version number of a class given a date}
\usage{
get_version(classification = NULL, date = NULL, family = NULL, klassNr = FALSE)

GetVersion(klass = NULL, date = NULL, family = NULL, klassNr = FALSE)
}
\arguments{
\item{classification}{Classification number}

\item{date}{Date for version to be valid}

\item{family}{Family ID number if a list of version number for all classes is desired}

\item{klassNr}{True/False for whether to output classification numbers. Default = FALSE}

\item{klass}{Deprecated; use `classification` instead.}
}
\value{
Number, vector or data frame with version numbers and calssification numbers if specified.
}
\description{
Get version number of a class given a date
}
\examples{
get_version(7)
}
