% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klass_list.R
\name{GetBaseUrl}
\alias{GetBaseUrl}
\title{Get base address for fetching classifications from}
\usage{
GetBaseUrl()
}
\value{
String with address
}
\description{
Get base address for fetching classifications from
}
\keyword{internal}
