% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klass_list.R
\name{GetNums}
\alias{GetNums}
\title{Get target ID numbers from Url}
\usage{
GetNums(x)
}
\arguments{
\item{x}{Url address}
}
\value{
Number
}
\description{
Get target ID numbers from Url
}
\keyword{internal}
