% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klass_list.R
\name{GetUrl}
\alias{GetUrl}
\title{Get json file from Url}
\usage{
GetUrl(url)
}
\arguments{
\item{url}{String url address}
}
\value{
text in json format
}
\description{
Get json file from Url
}
\keyword{internal}
