% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hent_data.R
\name{GetUrl2}
\alias{GetUrl2}
\title{Get json file from Url - alternative version}
\usage{
GetUrl2(url, check = TRUE)
}
\arguments{
\item{url}{String url address}

\item{check}{Logical parameter on whether to check if the url exists}
}
\value{
text in json format
}
\description{
Get json file from Url - alternative version
}
\keyword{internal}
