% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateKlass-graph-navigate.R
\name{is_split}
\alias{is_split}
\title{Given a graph and a node, determine if the node is a split code.}
\usage{
is_split(graph, node)
}
\arguments{
\item{graph}{A graph generated by \code{\link{klass_graph}}.}

\item{node}{A node as returned by \code{\link{klass_node}} or
\code{\link[igraph]{V}}.}
}
\value{
\code{TRUE} if the node is split, otherwise \code{FALSE}.
}
\description{
Given a graph and a node, determine if the node is a split code.
}
\details{
The function will attempt to reconcile nodes that have split and
  then later merged again. A node is considered to be split if there is more
  than one node that does not itself have children (i.e. nodes at the end of
  a sequence of changes) that can be reached from \code{node}
}
\keyword{internal}
