% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{klass_131_graph}
\alias{klass_131_graph}
\title{Test Graph data for municipalities in 2024}
\format{
An object of class \code{igraph} of length 2000.
}
\usage{
klass_131_graph
}
\description{
A nested list of graph data for using in testing
}
\keyword{datasets}
