% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLevel.R
\name{levelCheck}
\alias{levelCheck}
\title{Check which level the input data/vector is}
\usage{
levelCheck(x, klass_data)
}
\arguments{
\item{x}{- vector of character}

\item{klass_data}{- - the right formatting to the classification levels}
}
\value{
The hierarchical level of the input data is returned.
}
\description{
The input level is calculated as the mode (most common) hierarchical
level in the input vector.
}
\keyword{internal}
