% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Levels.R
\name{Levels}
\alias{Levels}
\title{Title}
\usage{
Levels(input_level, output_level, klass_data)
}
\arguments{
\item{input_level}{Classification level from the original dataset.}

\item{output_level}{Classification level for which the codes and names are desired by the user.}

\item{klass_data}{Classification file retrieved from KLASS.}
}
\value{
Data frame with the input and extra desired classification levels. Includes codes and names for each level.
}
\description{
Title
}
\keyword{internal}
