% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klass_list.R
\name{list_family}
\alias{list_family}
\alias{ListFamily}
\title{Classification family list
Print a list of all families and the number of classifications in each}
\usage{
list_family(family = NULL, codelists = FALSE, language = "nb")

ListFamily(family = NULL, codelists = FALSE, language = "nb")
}
\arguments{
\item{family}{Input family ID number to get a list of classifications in that family}

\item{codelists}{True/False for whether to include codelists. Default = FALSE}

\item{language}{Two letter string for the requested language output. Default is Bokmål ("nb"). Nynorsk ("nn") and English ("en").}
}
\value{
dataset containing a list of families
}
\description{
Classification family list
Print a list of all families and the number of classifications in each
}
\examples{
list_family(family = 1)
}
