% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klass_list.R
\name{list_klass}
\alias{list_klass}
\alias{ListKlass}
\title{Classification list
Get a full list of all classifications and codelists}
\usage{
list_klass(codelists = FALSE, language = "nb")

ListKlass(codelists = FALSE, language = "nb")
}
\arguments{
\item{codelists}{True/False for whether to include codelists. Default = FALSE}

\item{language}{Two letter string for the requested language output. Default is Bokmål ("nb"). Nynorsk ("nn") and English ("en").}
}
\value{
A data frame containing a full list of classifications. The data frame includes the classification name, number, family and type.
}
\description{
Classification list
Get a full list of all classifications and codelists
}
\examples{
head(list_klass(codelists = TRUE))
}
