% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Klass_list.R
\name{search_klass}
\alias{search_klass}
\alias{SearchKlass}
\title{Search Klass}
\usage{
search_klass(query, codelists = FALSE, size = 20)

SearchKlass(query, codelists = FALSE, size = 20)
}
\arguments{
\item{query}{String with key word to search for}

\item{codelists}{True/False for whether to include codelists. Default = FALSE}

\item{size}{The number of results to show. Default = 20.}
}
\value{
Data frame of possible classifications that match the query
}
\description{
Search Klass
}
\examples{
search_klass("occupation")
}
