% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hent_data.R
\name{stop_quietly}
\alias{stop_quietly}
\title{Stop quietly function
Stop from a function without an error. Used for stopping when no internet}
\usage{
stop_quietly()
}
\description{
Stop quietly function
Stop from a function without an error. Used for stopping when no internet
}
\keyword{internal}
